lineoperations: split files into fns and prefs
Move free lo_info to prefs file.
