+++ lineoperations/src/lineoperations.c
@@ -20,11 +20,14 @@
 */
 
 
-#include "config.h"
+#ifdef HAVE_CONFIG_H
+    #include "config.h" /* for the gettext domain */
+#endif
 
 #include <geanyplugin.h>
 #include "Scintilla.h"
-#include "linefunctions.h"
+#include "lo_fns.h"
+#include "lo_prefs.h"
 
 
 static GtkWidget *main_menu_item = NULL;
@@ -48,22 +51,6 @@ struct lo_menu_item
 	GCallback cb_activate;
 	gpointer cb_data;
 };
-
-
-static struct
-{
-	/* general settings */
-	gchar *config_file;
-	gboolean use_collation_compare;
-} AddonsInfo;
-static AddonsInfo *lo_info = NULL;
-
-
-static struct
-{
-    GtkWidget *collation_cb;
-}
-static config_widgets;
 
 
 typedef void (*CB_USER_FUNCTION)(GtkMenuItem *menuitem, gpointer gdata);
@@ -238,7 +225,6 @@ action_indir_manip_item(GtkMenuItem *menuitem, gpointer gdata)
 }
 
 
-
 /*
  * Menu action for functions with direct scintilla manipulation
  * e.g. no need for **lines array, *new_file...
@@ -264,69 +250,6 @@ action_sci_manip_item(GtkMenuItem *menuitem, gpointer gdata)
 	user_indicate(doc->editor, lines_affected, &sel);
 
 	sci_end_undo_action(doc->editor->sci);
-}
-
-
-/* handle button presses in the preferences dialog box */
-static void
-lo_configure_response_cb(GtkDialog *dialog, gint response, gpointer user_data)
-{
-    if (response == GTK_RESPONSE_OK || response == GTK_RESPONSE_APPLY)
-    {
-		GKeyFile *config = g_key_file_new();
-		gchar *config_dir = g_path_get_dirname(lo_info->config_file);
-		gchar *data;
-
-        /* Grabbing options that has been set */
-        lo_info->use_collation_compare =
-            gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(config_widgets.collation_cb));
-
-        /* Write preference to file */
-        g_key_file_load_from_file(config, lo_info->config_file, G_KEY_FILE_NONE, NULL);
-
-        g_key_file_set_boolean(config, "general", "use_collation_compare",
-            lo_info->use_collation_compare);
-
-        if (!g_file_test(config_dir, G_FILE_TEST_IS_DIR)
-            && utils_mkdir(config_dir, TRUE) != 0)
-        {
-            dialogs_show_msgbox(GTK_MESSAGE_ERROR,
-                _("Plugin configuration directory could not be created."));
-        }
-        else
-        {
-            /* write config to file */
-            data = g_key_file_to_data(config, NULL, NULL);
-            utils_write_file(lo_info->config_file, data);
-            g_free(data);
-        }
-
-        g_free(config_dir);
-        g_key_file_free(config);
-    }
-}
-
-
-/* Configure the preferences GUI and callbacks */
-static GtkWidget *
-lo_configure(G_GNUC_UNUSED GeanyPlugin *plugin, GtkDialog *dialog, G_GNUC_UNUSED gpointer pdata)
-{
-    GtkWidget *vbox;
-    vbox = gtk_vbox_new(FALSE, 6);
-
-    config_widgets.collation_cb = gtk_check_button_new_with_label(
-        _("Use collation based string compare"));
-
-    gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON(config_widgets.collation_cb),
-        lo_info->use_collation_compare);
-
-    gtk_box_pack_start(GTK_BOX(vbox), config_widgets.collation_cb, FALSE, FALSE, 2);
-
-
-    gtk_widget_show_all(vbox);
-    g_signal_connect(dialog, "response", G_CALLBACK(lo_configure_response_cb), NULL);
-
-    return vbox;
 }
 
 
@@ -363,28 +286,6 @@ static void lo_keybinding_callback(guint key_id)
 	g_return_if_fail(key_id < G_N_ELEMENTS(menu_items));
 	cb_activate = (CB_USER_FUNCTION)menu_items[key_id].cb_activate;
 	cb_activate(NULL, menu_items[key_id].cb_data);
-}
-
-
-/* Initialize preferences */
-static void
-lo_init_prefs(GeanyPlugin *plugin)
-{
-	GeanyData *geany_data = plugin->geany_data;
-	GKeyFile *config = g_key_file_new();
-
-	/* load preferences from file into lo_info */
-	lo_info = g_new0(AddonsInfo, 1);
-	lo_info->config_file = g_strconcat(geany->app->configdir,
-		G_DIR_SEPARATOR_S, "plugins", G_DIR_SEPARATOR_S,
-		"lineoperations", G_DIR_SEPARATOR_S, "general.conf", NULL);
-
-	gboolean test = g_key_file_load_from_file(config, lo_info->config_file, G_KEY_FILE_NONE, NULL);
-
-	lo_info->use_collation_compare = utils_get_setting_boolean(config,
-		"general", "use_collation_compare", FALSE);
-
-	g_key_file_free(config);
 }
 
 
@@ -458,8 +359,7 @@ static void
 lo_cleanup(GeanyPlugin *plugin, gpointer pdata)
 {
 	gtk_widget_destroy(main_menu_item);
-	g_free(lo_info->config_file);
-	g_free(lo_info);
+	lo_free_info();
 }
